+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 ### Changed
 
 - Nécessite SPIP 4.2.0-dev minimum
+
+### Fixed
+
+- #4855 Ne pas nécessiter la présence d’un htaccess pour traiter l’url d’api de statistiques